gym-app/backend/requirements.txt
django>=4.2
psycopg2-binary
Pillow>=10.0
djangorestframework>=3.15.1
django-filter>=24.1
